+++ src/rpm/package.rs
@@ -324,6 +324,12 @@ impl RPMPackageMetadata {
     }
 
     #[inline]
+    pub fn get_cookie(&self) -> Result<&str, RPMError> {
+        self.header
+            .get_entry_data_as_string(IndexTag::RPMTAG_COOKIE)
+    }
+
+    #[inline]
     pub fn get_source_rpm(&self) -> Result<&str, RPMError> {
         self.header
             .get_entry_data_as_string(IndexTag::RPMTAG_SOURCERPM)